requirements.txt
chromadb==0.6.0
google_cloud_aiplatform==1.120.0
langchain_google_vertexai==3.0.0
openai==2.5.0
protobuf==6.33.0
pytest==8.4.2
python-dotenv==1.1.1
vertexai==1.71.1
